+++ planning/motion_velocity_planner/autoware_motion_velocity_obstacle_velocity_limiter_module/src/obstacle_velocity_limiter.hpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLE_VELOCITY_LIMITER_HPP_
-#define AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLE_VELOCITY_LIMITER_HPP_
+#ifndef OBSTACLE_VELOCITY_LIMITER_HPP_
+#define OBSTACLE_VELOCITY_LIMITER_HPP_
 
-#include "obstacle_velocity_limiter/obstacles.hpp"
-#include "obstacle_velocity_limiter/parameters.hpp"
-#include "obstacle_velocity_limiter/types.hpp"
+#include "obstacles.hpp"
+#include "parameters.hpp"
+#include "types.hpp"
 
 #include <autoware_perception_msgs/msg/predicted_objects.hpp>
 #include <nav_msgs/msg/occupancy_grid.hpp>
@@ -129,4 +129,4 @@ Trajectory copyDownsampledVelocity(
   const int factor);
 }  // namespace obstacle_velocity_limiter
 
-#endif  // AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLE_VELOCITY_LIMITER_HPP_
+#endif  // OBSTACLE_VELOCITY_LIMITER_HPP_
