+++ planning/motion_velocity_planner/autoware_motion_velocity_obstacle_velocity_limiter_module/src/obstacle_velocity_limiter_node.hpp
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLE_VELOCITY_LIMITER_NODE_HPP_
-#define AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLE_VELOCITY_LIMITER_NODE_HPP_
+#ifndef OBSTACLE_VELOCITY_LIMITER_NODE_HPP_
+#define OBSTACLE_VELOCITY_LIMITER_NODE_HPP_
 
-#include "obstacle_velocity_limiter/obstacles.hpp"
-#include "obstacle_velocity_limiter/parameters.hpp"
-#include "obstacle_velocity_limiter/types.hpp"
+#include "obstacles.hpp"
+#include "parameters.hpp"
 #include "tier4_autoware_utils/ros/logger_level_configure.hpp"
+#include "types.hpp"
 
 #include <rclcpp/rclcpp.hpp>
 #include <tier4_autoware_utils/ros/published_time_publisher.hpp>
@@ -110,4 +110,4 @@ private:
 };
 }  // namespace obstacle_velocity_limiter
 
-#endif  // AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLE_VELOCITY_LIMITER_NODE_HPP_
+#endif  // OBSTACLE_VELOCITY_LIMITER_NODE_HPP_
