+++ planning/motion_velocity_planner/autoware_motion_velocity_obstacle_velocity_limiter_module/src/obstacles.hpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLES_HPP_
-#define AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLES_HPP_
+#ifndef OBSTACLES_HPP_
+#define OBSTACLES_HPP_
 
-#include "obstacle_velocity_limiter/parameters.hpp"
-#include "obstacle_velocity_limiter/types.hpp"
+#include "parameters.hpp"
+#include "types.hpp"
 
 #include <tier4_autoware_utils/ros/transform_listener.hpp>
 
@@ -185,4 +185,4 @@ void addSensorObstacles(
 /// @return obstacles that are inside the positive mask and outside of the negative masks
 Obstacles filterObstacles(const Obstacles & obstacles, const ObstacleMasks & masks);
 }  // namespace obstacle_velocity_limiter
-#endif  // AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__OBSTACLES_HPP_
+#endif  // OBSTACLES_HPP_
