+++ planning/motion_velocity_planner/autoware_motion_velocity_obstacle_velocity_limiter_module/src/parameters.hpp
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__PARAMETERS_HPP_
-#define AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__PARAMETERS_HPP_
+#ifndef PARAMETERS_HPP_
+#define PARAMETERS_HPP_
 
-#include <obstacle_velocity_limiter/types.hpp>
+#include "types.hpp"
+
 #include <rclcpp/node.hpp>
 
 #include <tf2/utils.h>
@@ -239,4 +240,4 @@ struct PreprocessingParameters
 };
 
 }  // namespace obstacle_velocity_limiter
-#endif  // AUTOWARE_MOTION_VELOCITY_OBSTACLE_VELOCITY_LIMITER_MODULE__PARAMETERS_HPP_
+#endif  // PARAMETERS_HPP_
